% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{extract_distribution_info}
\alias{extract_distribution_info}
\title{Extract distribution and growth rate from fitting grid}
\usage{
extract_distribution_info(fitting_grid)
}
\arguments{
\item{fitting_grid}{Single row from fitting grid with distribution and
growth_rate columns}
}
\value{
List with distribution and growth_rate
}
\description{
Extract distribution and growth rate from fitting grid
}
