% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_primarycensored_mle}
\alias{fit_primarycensored_mle}
\title{Fit primarycensored MLE model using fitdistrplus}
\usage{
fit_primarycensored_mle(fitting_grid)
}
\arguments{
\item{fitting_grid}{Single row from fitting grid}
}
\value{
Data frame with parameter estimates and diagnostics
}
\description{
Fit primarycensored MLE model using fitdistrplus
}
