% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_ward}
\alias{fit_ward}
\title{Fit Ward et al. latent variable model}
\usage{
fit_ward(fitting_grid, stan_settings, model = NULL)
}
\arguments{
\item{fitting_grid}{Single row from fitting grid}

\item{stan_settings}{List of Stan sampling settings}

\item{model}{Optional pre-compiled Stan model (will compile if NULL)}
}
\value{
Data frame with parameter estimates and diagnostics
}
\description{
Fit Ward et al. latent variable model
}
