% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate a Dataset of Censored Delays}
\usage{
generate_dataset(
  rdist,
  obs_time,
  pw_range = 1:4,
  sw_range = 1:4,
  n = 10000,
  ...
)
}
\arguments{
\item{rdist}{A random number generating function that will be used to
generate delays}

\item{obs_time}{A single value or vector of observation times. If a vector is
provided, it will be sampled randomly for each observation.}

\item{pw_range}{Range of possible values for primary window, sampled randomly
(default: 1:4)}

\item{sw_range}{Range of possible values for secondary window, sampled
randomly (default: 1:4)}

\item{n}{Number of samples to generate (default: 10000)}

\item{...}{Additional parameters passed to the rdist function}
}
\value{
A data frame containing:
\item{pwindow}{Primary window size for each observation}
\item{obs_time}{Observation time for each sample}
\item{observed_delay}{The observed (censored) delay}
\item{observed_delay_upper}{Upper bound of the delay, minimum of obs_time
and observed_delay + swindow}
}
\description{
This function generates a simulated dataset of censored delays using the
specified random distribution and parameters.
}
\details{
The function generates censored delay data using the primarycensored
package's rpcens function. For each observation, it randomly samples primary
window sizes, secondary window sizes, and observation times from the provided
ranges.
}
\examples{
# Generate dataset with exponential delays
library(primarycensoredpaper)
data <- generate_dataset(rexp, obs_time = 1:10, rate = 0.1)

}
