% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_distribution_id}
\alias{get_distribution_id}
\title{Get distribution ID for Stan models}
\usage{
get_distribution_id(distribution)
}
\arguments{
\item{distribution}{Character string: "gamma" or "lognormal"}
}
\value{
Integer distribution ID for Stan (1=lognormal, 2=gamma, matches primarycensored)
}
\description{
Get distribution ID for Stan models
}
