% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_tools.R
\name{get_dprimary}
\alias{get_dprimary}
\title{Get primary density function based on growth rate}
\usage{
get_dprimary(growth_rate)
}
\arguments{
\item{growth_rate}{Growth rate parameter}
}
\value{
Function for primary distribution density (dunif if
growth_rate is 0, dexpgrowth otherwise)
}
\description{
Get primary density function based on growth rate
}
