% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_tools.R
\name{get_dprimary_args}
\alias{get_dprimary_args}
\title{Get primary density arguments based on growth rate}
\usage{
get_dprimary_args(growth_rate)
}
\arguments{
\item{growth_rate}{Growth rate parameter}
}
\value{
List of arguments for the primary distribution density
}
\description{
Get primary density arguments based on growth rate
}
