% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{get_param_names}
\alias{get_param_names}
\title{Get parameter names for distribution}
\usage{
get_param_names(distribution)
}
\arguments{
\item{distribution}{Character string: "gamma" or "lognormal"}
}
\value{
Character vector of parameter names
}
\description{
Get parameter names for distribution
}
