% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_tools.R
\name{get_primary_args}
\alias{get_primary_args}
\title{Get primary distribution arguments based on growth rate}
\usage{
get_primary_args(growth_rate)
}
\arguments{
\item{growth_rate}{Growth rate parameter}
}
\value{
List of arguments for the primary distribution
}
\description{
Get primary distribution arguments based on growth rate
}
