% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_tools.R
\name{get_primary_dist}
\alias{get_primary_dist}
\title{Get primary distribution function based on growth rate}
\usage{
get_primary_dist(growth_rate)
}
\arguments{
\item{growth_rate}{Growth rate parameter}
}
\value{
Function for primary distribution (uniform if growth_rate is 0,
exponential growth otherwise)
}
\description{
Get primary distribution function based on growth rate
}
