% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{get_r_distribution_name}
\alias{get_r_distribution_name}
\title{Map distribution names to R distribution function names}
\usage{
get_r_distribution_name(distribution)
}
\arguments{
\item{distribution}{Character string: "gamma" or "lognormal"}
}
\value{
Character string of R distribution name
}
\description{
Map distribution names to R distribution function names
}
