% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{get_relative_obs_time}
\alias{get_relative_obs_time}
\title{Get relative observation time from truncation scenario}
\usage{
get_relative_obs_time(truncation)
}
\arguments{
\item{truncation}{Character string: truncation scenario name}
}
\value{
Numeric value for relative observation time
}
\description{
Get relative observation time from truncation scenario
}
