% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_tools.R
\name{get_rprimary}
\alias{get_rprimary}
\title{Get primary random generation function based on growth rate}
\usage{
get_rprimary(growth_rate)
}
\arguments{
\item{growth_rate}{Growth rate parameter}
}
\value{
Function for primary distribution random generation (runif if
growth_rate is 0, rexpgrowth otherwise)
}
\description{
Get primary random generation function based on growth rate
}
