% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_tools.R
\name{get_rprimary_args}
\alias{get_rprimary_args}
\title{Get primary random generation arguments based on growth rate}
\usage{
get_rprimary_args(growth_rate)
}
\arguments{
\item{growth_rate}{Growth rate parameter}
}
\value{
List of arguments for the primary distribution random generation
}
\description{
Get primary random generation arguments based on growth rate
}
