% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{get_start_values}
\alias{get_start_values}
\title{Get starting values for distribution fitting}
\usage{
get_start_values(distribution)
}
\arguments{
\item{distribution}{Character string: "gamma" or "lognormal"}
}
\value{
List of starting parameter values
}
\description{
Get starting values for distribution fitting
}
