% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{prepare_ward_stan_data}
\alias{prepare_ward_stan_data}
\title{Prepare Ward-specific Stan data from shared model inputs}
\usage{
prepare_ward_stan_data(sampled_data, shared_inputs, bounds_priors)
}
\arguments{
\item{sampled_data}{Data frame with delay observations and censoring windows}

\item{shared_inputs}{Output from prepare_shared_model_inputs}

\item{bounds_priors}{Output from get_shared_prior_settings}
}
\value{
List of Stan data for Ward model
}
\description{
Takes the output from prepare_shared_model_inputs and adds Ward-specific
data requirements (censoring windows and observation times).
}
