% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_data}
\alias{save_data}
\title{Save data to the results directory}
\usage{
save_data(data, filename)
}
\arguments{
\item{data}{The data object to save}

\item{filename}{The filename (without path)}
}
\description{
Save data to the results directory
}
