% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot to the figures directory}
\usage{
save_plot(plot, filename, width = 8, height = 6, dpi = 300)
}
\arguments{
\item{plot}{The plot object to save}

\item{filename}{The filename (without path)}

\item{width}{Width in inches}

\item{height}{Height in inches}

\item{dpi}{Resolution in dots per inch}
}
\description{
Save a plot to the figures directory
}
