% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_tools.R
\name{setup_pmf_inputs}
\alias{setup_pmf_inputs}
\title{Setup PMF calculation inputs}
\usage{
setup_pmf_inputs(scenarios, distributions, growth_rate, is_numerical = FALSE)
}
\arguments{
\item{scenarios}{Scenario data frame row}

\item{distributions}{Distribution data frame}

\item{growth_rate}{Growth rate parameter}

\item{is_numerical}{Logical indicating if numerical integration should be
used}
}
\value{
List with calculation inputs
}
\description{
Setup PMF calculation inputs
}
