% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_ebola_windows}
\alias{summarise_ebola_windows}
\title{Summarise Ebola delay data by window and analysis type}
\usage{
summarise_ebola_windows(ebola_delay_data)
}
\arguments{
\item{ebola_delay_data}{A data frame with rows containing metadata and
nested delay data}
}
\value{
A data frame with summary statistics for each window/analysis
}
\description{
Summarise Ebola delay data by window and analysis type
}
