% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-utils.R
\name{get_start_values}
\alias{get_start_values}
\title{Get starting values for distribution fitting}
\usage{
get_start_values(distribution, data = NULL)
}
\arguments{
\item{distribution}{Character string: "gamma" or "lognormal"}

\item{data}{Optional data to compute method-of-moments starting values}
}
\value{
List of starting parameter values
}
\description{
Get starting values for distribution fitting
}
